requirements.txt
numpy==1.24.3
torch==2.0.1+cu117
torchvision==0.15.2+cu117
Pillow==9.5.0
torchmetrics==0.11.4
tqdm==4.65.0
wandb==0.15.2